add mnn for different layers
